Move selected functions to top-level, create _ops for specific cross-subject functions
